Refactor : 도메인 계층 Repository 삭제
