+++ src/main/java/com/driven/dm/cart/application/service/CartService.java
@@ -15,8 +15,7 @@ import com.driven.dm.cart.presentation.dto.response.UserCartSummaryDto;
 import com.driven.dm.cart.presentation.dto.response.UserCartsResponse;
 import com.driven.dm.global.exception.AppException;
 import com.driven.dm.menu.domain.entity.Menu;
-import com.driven.dm.menu.domain.repository.MenuRepository;
-import com.driven.dm.menu.infrastructure.repository.MenuJpaRepository;
+import com.driven.dm.menu.infrastructure.repository.MenuRepository;
 import com.driven.dm.shop.application.exception.ShopErrorCode;
 import com.driven.dm.shop.domain.entity.Shop;
 import com.driven.dm.shop.domain.repository.ShopRepository;
@@ -60,7 +59,7 @@ public class CartService {
             throw new AppException(ShopErrorCode.SHOP_NOT_FOUND);
         }
 
-        Menu menu = menuRepository.selectMenu(req.getMenuId())
+        Menu menu = menuRepository.findById(req.getMenuId())
             .orElseThrow(() -> new AppException(CartErrorCode.INVALID_MENU));
 
         Cart cart = cartRepository
