+++ src/main/java/com/driven/dm/shop/presentation/dto/request/ShopUpdateRequest.java
@@ -23,6 +23,7 @@ public class ShopUpdateRequest {
     @Pattern(regexp = "OPEN|CLOSED", message = "가게 상태는 OPEN 또는 CLOSED만 가능합니다.")
     private String status;
 
+    @NotBlank
     @Pattern(regexp = "KOREAN|CHINESE|SNACK|CHICKEN|PIZZA",
         message = "카테고리는 KOREAN, CHINESE, SNACK, CHICKEN, PIZZA만 가능합니다.")
     private String category;
