+++ src/test/java/com/driven/dm/order/application/service/OrderServiceTest.java
@@ -7,7 +7,7 @@ import static org.mockito.Mockito.mock;
 
 import com.driven.dm.global.exception.AppException;
 import com.driven.dm.menu.domain.entity.Menu;
-import com.driven.dm.menu.domain.repository.MenuRepository;
+import com.driven.dm.menu.infrastructure.repository.MenuRepository;
 import com.driven.dm.order.application.exception.OrderErrorCode;
 import com.driven.dm.order.domain.entity.Order;
 import com.driven.dm.order.domain.entity.OrderStatus;
